Notice disconnects from ot-daemon. This allows us to automatically reconnect if the daemon crashes, or to automatically connect if srp-mdns-proxy starts before the daemon.
Store the ULA prefix in stable storage in /var/lib/openthread. This means that if srp-mdns-proxy is restarted, we don't suddenly get a different prefix.

Fix ifconfig command so that when we need to offer a ULA prefix on the infrastructure link, it's configured correctly on the interface. Subproc code also had some bugs because of API changes in macos-ioloop.c that hadn't been properly ported to ioloop.c yet.

Various Makefile fixes to make installing work correctly.

Use mbedtls package, not mbedtls from /usr/local

Address sanitizer support for gcc/linux