This pull request adds an implementation of SRP serial number support to both srp-client and srp-mdns-proxy.
The serial number is passed in an edns0 option, which is temporarily allocated for site-local use, but we need to ask for an early assignment as soon as possible so that this can go into ADK.

The edns0 option parsing in the client has been improved to the degree that it can now parse more than one option, in arbitrary order.

While working on this, I discovered several bugs in the open source ioloop.c code: first, the subprocess code was not properly refcounting in the case that the output was being captured. Second, the dnssd_txn_* support was not closing the transaction when the dnssd_txn_t object was released. Fixing this involved changing srp-mdns-proxy to actually cancel the transaction before freeing it. There was also a potential double free in srp-mdns-proxy when freeing the transaction objects, which didn't actually cause a double free because the transactions weren't being freed correctly in the first place. This has been cleaned up.

The code for adding an interface route allocated a global subproc_t, which doesn't work if there's more than one interface and the routes are installed at the same time. This was corrected by making the route add subprocess per-interface.

There was a lot of code in route.c that released pointers stored in globals and didn't zero the globals when subproc callbacks were called. Most of this code is never executed, and the pointers would never be double-freed, but it seemed best to clean it up since I'd noticed it.