This PR fixes the initial problem that showed up under ASAN, and a couple of other problems that were not as obvious. First, the code was a little too clever about detecting that the wakeup list had been modified, with the result that we could dereference the wakeup object after it had been freed by the callback. This probably didn't cause the reported crash.
The second issue was that host_dereference() would free the dnssd_txn object hanging off the host object, and this would trigger an asynchronous call to the txn_finalize function, which would then write to the host object, which had already been freed. This doesn't happen on MacOS/tvOS because the callback is called synchronously, before the call to dnssd_txn_dereference() returns.

A third issue which likely wouldn't ever happen other than under the debugger is that the link route configuration subprocess code could run a second time while the first invocation of the subprocess was outstanding; in this case, the callback would wind up dereferencing the wrong pointer. The fix is to just not invoke a second subprocess if one is outstanding.

A fourth issue in srp-ioloop was that the wakeup associated with the udp_context would not be canceled or freed when the udp_context was finalized. Ordinarily this had no effect, but if the network state changed, triggering a new context, when the wakeup fired we'd get a crash. The fix is to cancel the wakeup and release it.

Finally, the code in srp-client for parsing EDNS0 options had a bug that would cause it to try to decode an option past the end of a buffer. This would always fail, but it caused a crash under the address sanitizer. This is now fixed.